revert entity id type

+++ Engine/Scene/Core/Entity.cpp
@@ -1,12 +1,12 @@
 #include "Entity.hpp"
 
-std::atomic<uint32_t> Entity::id_counter = 0;
+std::atomic<uint64_t> Entity::id_counter = 0;
 
 Entity::Entity() :
     uid(id_counter++)
 {}
 
-uint32_t Entity::getUid() const
+uint64_t Entity::getUid() const
 {
 	return uid;
 }
